api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
bcrypt
pydantic-settings

alembic
python-jose[cryptography]
python-multipart
python-dotenv

redis
